pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ProductScraper"
version = "0.1.0"
description = "A project to scrape product information."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "beautifulsoup4>=4.10.0",
    "httpx>=0.25.0",
    "openpyxl>=3.1.0",
    "pandas>=2.3.3",
    "Pillow>=12.0.0",
    "psutil>=5.9.0",
    "PyQt6>=6.10.0",
    "pyqt6-sip>=13.10.2",
    "PyQt6-WebEngine>=6.10.0",
    "pytest>=9.0.1",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "tenacity>=8.2.0",
    "undetected-chromedriver>=3.5.5",
    "fake-useragent>=1.4.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "ollama"
]

[project.optional-dependencies]
dev = [
    "flask==3.1.2",
    "pytest-asyncio",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests",
    "types-PyYAML",
    "pyinstaller>=6.0.0"
]

[project.scripts]
productscraper = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
norecursedirs = ["tests/unit/test_scrapers.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "performance: marks performance tests",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
    "security: marks security tests",
]
addopts = "--strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
